Add Flyway DB migrations.
The StorageManager creates an actor to perform the migrations. If a retryable
failure occurs, that actor is supervised by a Backoff supervisor that will
keep trying to start the migrations. Fatal errors are escalated to the guardian,
which will shutdown the actor system. Currently this will _not_ bring down the
whole app, but it probably should.